requirements.txt
flask
google-generativeai
sqlalchemy
pdf2image
pillow
python-dotenv
gunicorn
pytest
requests
